Fix chunked decoding on temporary read error [RT#74431]
If the first read after a chunk header returns EAGAIN or EINTR, the
chunked state stored in the client object doesn't get updated, so the
next call to ->read_entity_body tries to read the chunk header again,
and gives a "Missing newline after chunk data" error.

This fixes it by updating ${*$self}{'http_chunked'} immediately after
the chunk header has been read, so that the state persist to the next
call.