Ignore automatically generated directories
When installing the dependencies with `cpm` I noticed that the `local/`
directory is created automatically and doesn't need to be checked in to
the repository, hence it can be ignored so that the working directory
remains clean.  The same is for the `cover_db/` directory which is
created when generating coverage information.