Add initial set of contribution guidelines
This change documents some of the steps required in order to contribute
to `Net::HTTP` in the hope that it helps newcomers to the project create
their first submission(s) and is submitted in response to a discussion
in #76.  It includes a quick start guide for setting up a `perlbrew`
environment as well as how to install `Dist::Zilla` and how to use it to
install upstream dependencies.  The document also describes the steps
necessary to create a pull request from a feature branch in a user's
forked repository.